+++ CHANGELOG.md
@@ -16,12 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * The MQTT settings validation handler now receives the already updated new settings
   and the old settings (instead of the not updated current settings and the new settings).
 * The MQTT client has been split into its own `miniconf_mqtt` crate.
+* The attribute syntax has changed from `#[tree(depth(1))]` to `#[tree(depth=1)]`.
+* The default depth is `0`, also in the case where a `#[tree()]` without `depth` has been specified.
 
 ### Added
 
 * Python lib: Support for clearing a retained setting
 * Python CLI: get() support
 * `TreeKey::iter_indices()` and `iter_indices_unchecked()`
+* Support for validation callbacks and errors to check deserialized data.
 
 ## [0.9.0](https://github.com/quartiq/miniconf/compare/v0.8.0...v0.9.0)